+++ src/server/getters-setters.ts
@@ -112,9 +112,9 @@ export class GettersSetters {
     console.log(chalk.gray("Done."));
   }
 
-  setDb(db?: Params.Db, { reverse = this.config.reverse }: GetValidDbOptions = {}) {
+  setDb(db?: Params.Db, { reverse = this.config.reverse, mode = this.config.mode }: GetValidDbOptions = {}) {
     console.log("\n" + chalk.gray("Setting Db..."));
-    this.#db = getValidDb(db, this.#injectors, this.#config.root, { reverse });
+    this.#db = getValidDb(db, this.#injectors, this.#config.root, { reverse, mode });
     this.initialDb = _.cloneDeep(this.#db);
     console.log(chalk.gray("Done."));
   }