+++ src/server/types/user.types.ts
@@ -4,7 +4,12 @@ import { Default_Middlewares, Global_Middlweares, HarMiddleware, KibanaMiddlewar
 import * as ValidTypes from './valid.types';
 
 export type Config = Partial<ValidTypes.Config>;
-export type Db = { [key: string]: RouteConfig } | { [key: string]: Omit<Object, "_config"> | any[] | string };
+export type Db =
+  { [key: string]: RouteConfig }
+  | { [key: string]: express.RequestHandler }
+  | { [key: string]: Omit<Object, "_config"> }
+  | { [key: string]: any[] }
+  | { [key: string]: string };
 export type Injectors = InjectorConfig[];
 export type Middlewares = Partial<Default_Middlewares & Global_Middlweares & HarMiddleware & KibanaMiddleware & User_Middlweares>
 export type Rewriters = RoutePairs;