+++ src/server/types/valid.types.ts
@@ -1,5 +1,5 @@
 import { AxiosRequestConfig } from 'axios';
-import { Default_Middlewares, Global_Middlweares, HarMiddleware, KibanaMiddleware, RoutePairs, User_Middlweares } from './common.types';
+import { DbMode, Default_Middlewares, Global_Middlweares, HarMiddleware, KibanaMiddleware, RoutePairs, User_Middlweares } from './common.types';
 import * as UserTypes from "./user.types";
 
 export type Config = {
@@ -15,6 +15,7 @@ export type Config = {
   logger: boolean;
   readOnly: boolean;
   bodyParser: boolean;
+  mode: DbMode
 };
 export type Db = { [key: string]: RouteConfig }
 export type Injectors = InjectorConfig[];